refactor: move radio types into radio module
