+++ rigflow-server/src/lib.rs
@@ -5,6 +5,7 @@ pub mod events;
 pub mod net;
 pub mod plugin_api;
 pub mod plugins;
+pub mod radio;
 pub mod server;
 pub mod waterfall;
 pub mod source;
