+++ rigflow-server/src/server/mod.rs
@@ -1,6 +1,5 @@
 pub mod discovery;
 pub mod radio_manager;
 pub mod radio_api;
-pub mod radio_types;
 pub mod radio_worker;
 pub mod session;
